Manual update script for non-root AGH service
This is a manual update script for arm64 linux boxes running AGH as non-root in systemctl --service mode, using 'AmbientCapabilities=CAP_NET_BIND_SERVICE CAP_NET_RAW' to bind to Port 53 and other privileged ports. Auto-Update is disabled in the web-interface under these circumstances, so this script makes updating a little easier.

It downloads the latest tarball and checksums from github into /tmp/, performs a checksum match, makes a backup of the AdGuardHome.yml config file and data directory into the users $HOME directory, and extracts and installs the new AdGuardHome binary, before finally starting the AdGuardHome.service

!! Please note. I am running AGH on a 64-bit raspberrypi. In its current form it only downloads the linux_arm64 tarball. If you are running a different arch, you will need to change the TARBALL line to match your architecture as it is written in AGH's github repo Releases !!

Would be great if someone more advanced than me could edit this to automate a check for the architecture and download the relevant binary.